requirements.txt
allure-behave==2.6.3
allure-python-commons==2.6.3
attrs==19.1.0
behave==1.2.6
certifi==2019.3.9
chardet==3.0.4
Django==2.2.2
django-filer==1.5.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-polymorphic==2.0.3
djangorestframework==3.9.3
easy-thumbnails==2.6
ExifRead==2.1.2
idna==2.8
parse==1.12.0
parse-type==0.4.2
Pillow==6.0.0
#pluggy==0.11.0
pytz==2019.1
requests==2.21.0
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
Unidecode==1.0.23
urllib3==1.24.3
